feat: change carbon footprint plugin to environmental footprint plugin
Change the initial usage of the carbon footprint plugin in SimGrid to be able to use the extension that we are developing, the environmental footprint plugin. With this extension, we can define a energy grid composition for the host, and also define the carbon or/and water footprint of the respective sources.

Signed-off-by: Victor Lima <lima.p.victor@outlook.com>