fix: change the README for trace CSV format
The README in the folder data/ was describing a trace csv format to deal with the old carbon footprint plugin. With the enhancements being made, now the host can have it energy grid composition described and also the carbon or/and water footprint for each source in the grid. So the expected format of trace changed, and we have to update the description of it.

Signed-off-by: Victor Lima <lima.p.victor@outlook.com>